change relation between core classes and ros2 nodes
