+++ as2_aerial_platforms/as2_platform_crazyflie/src/crazyflie_node.cpp
@@ -40,7 +40,7 @@ int main(int argc, char* argv[]) {
   rclcpp::init(argc, argv);
   auto node = std::make_shared<CrazyfliePlatform>();
   node->preset_loop_frequency(70);
-  as2::spinLoop(node);
+  as2::spinLoop(node, node);
   RCLCPP_INFO(node->get_logger(), "Ending!");
 
   rclcpp::shutdown();
