+++ as2_core/src/node.cpp
@@ -34,19 +34,3 @@
  ********************************************************************************/
 
 #include "as2_core/node.hpp"
-
-std::string as2::Node::generate_global_name(const std::string &name) {
-  if (name.find("/") == 0) {
-    return name.substr(1);
-  } else {
-    return name;
-  }
-};
-
-std::string as2::Node::generate_local_name(const std::string &name) {
-  if (name.find("/") == 0) {
-    return this->get_name() + name;
-  } else {
-    return std::string(this->get_name()) + "/" + name;
-  }
-};
