+++ as2_core/tests/platform_state_machine_test.cpp
@@ -43,7 +43,10 @@
 
 TEST(PlatformStateMachineTest, Normal_functionality) {
   as2::Node test_node("test_node");
-  as2::PlatformStateMachine state_machine(&test_node);
+  as2::NodeWithInit node;
+  node.init(&test_node);
+  as2::PlatformStateMachine state_machine;
+  state_machine.start(&test_node,&node);
   EXPECT_EQ(state_machine.getState().state, as2_msgs::msg::PlatformStatus::DISARMED);
   state_machine.processEvent(as2_msgs::msg::PlatformStateMachineEvent::ARM);
   EXPECT_EQ(state_machine.getState().state, as2_msgs::msg::PlatformStatus::LANDED);
@@ -61,8 +64,11 @@ TEST(PlatformStateMachineTest, Normal_functionality) {
 
 TEST(PlatformStateMachineTest, EmergencyTest) {
   as2::Node test_node("test_node");
+  as2::NodeWithInit node;
+  node.init(&test_node);
   for (int i = -1; i < (int)as2_msgs::msg::PlatformStatus::LANDING; i++) {
-    as2::PlatformStateMachine state_machine(&test_node);
+    as2::PlatformStateMachine state_machine;
+    state_machine.start(&test_node,&node);
     state_machine.setState(i);
     state_machine.processEvent(as2_msgs::msg::PlatformStateMachineEvent::EMERGENCY);
     EXPECT_EQ(state_machine.getState().state, as2_msgs::msg::PlatformStatus::EMERGENCY);
@@ -71,7 +77,10 @@ TEST(PlatformStateMachineTest, EmergencyTest) {
 
 TEST(PlatformStateMachineTest, CorrectFSMActivation) {
   as2::Node test_node("test_node");
-  as2::PlatformStateMachine state_machine(&test_node);
+  as2::NodeWithInit node;
+  node.init(&test_node);
+  as2::PlatformStateMachine state_machine;
+  state_machine.start(&test_node,&node);
   EXPECT_EQ(state_machine.getState().state, as2_msgs::msg::PlatformStatus::DISARMED);
   state_machine.processEvent(as2_msgs::msg::PlatformStateMachineEvent::ARM);
   EXPECT_EQ(state_machine.getState().state, as2_msgs::msg::PlatformStatus::LANDED);
@@ -89,7 +98,10 @@ TEST(PlatformStateMachineTest, CorrectFSMActivation) {
 
 TEST(PlatformStateMachineTest, TakeOffTest) {
   as2::Node test_node("test_node");
-  as2::PlatformStateMachine state_machine(&test_node);
+  as2::NodeWithInit node;
+  node.init(&test_node);
+  as2::PlatformStateMachine state_machine;
+  state_machine.start(&test_node,&node);
   EXPECT_EQ(state_machine.getState().state, as2_msgs::msg::PlatformStatus::DISARMED);
   state_machine.processEvent(as2_msgs::msg::PlatformStateMachineEvent::ARM);
   EXPECT_EQ(state_machine.getState().state, as2_msgs::msg::PlatformStatus::LANDED);
